Move ts_to_symbol into util
