+++ creusot/src/translation/specification.rs
@@ -1,9 +1,8 @@
 use std::collections::HashMap;
 
-use crate::ctx::*;
 use crate::translation::specification;
+use crate::{ctx::*, util};
 use rustc_middle::ty::Attributes;
-use rustc_span::Symbol;
 use why3::declaration::Contract;
 use why3::mlcfg::Exp;
 
@@ -44,22 +43,6 @@ pub fn ensures_to_why<'tcx>(
     log::debug!("ensures clause {:?}", ens_id);
     let term = specification::typing::typecheck(ctx.tcx, ens_id.expect_local());
     lower_term_to_why3(ctx, names, ens_id, term)
-}
-
-use rustc_ast::{
-    token::TokenKind::Literal,
-    tokenstream::{TokenStream, TokenTree::*},
-};
-
-pub fn ts_to_symbol(ts: TokenStream) -> Option<Symbol> {
-    assert_eq!(ts.len(), 1);
-
-    if let Token(tok) = ts.trees().next().unwrap() {
-        if let Literal(lit) = tok.kind {
-            return Some(lit.symbol);
-        }
-    }
-    None
 }
 
 pub struct PreContract {
@@ -137,17 +120,17 @@ pub fn contract_of(tcx: TyCtxt, def_id: DefId) -> Result<PreContract, SpecAttrEr
 
         match attr.path.segments[2].ident.to_string().as_str() {
             "requires" => {
-                let req_name = ts_to_symbol(attr.args.inner_tokens()).ok_or(InvalidTokens)?;
+                let req_name = util::ts_to_symbol(attr.args.inner_tokens()).ok_or(InvalidTokens)?;
                 let req_id = tcx.get_diagnostic_item(req_name).ok_or(InvalidTokens)?;
                 contract.requires.push(req_id);
             }
             "ensures" => {
-                let ens_name = ts_to_symbol(attr.args.inner_tokens()).ok_or(InvalidTokens)?;
+                let ens_name = util::ts_to_symbol(attr.args.inner_tokens()).ok_or(InvalidTokens)?;
                 let ens_id = tcx.get_diagnostic_item(ens_name).ok_or(InvalidTokens)?;
                 contract.ensures.push(ens_id);
             }
             "variant" => {
-                let var_name = ts_to_symbol(attr.args.inner_tokens()).ok_or(InvalidTokens)?;
+                let var_name = util::ts_to_symbol(attr.args.inner_tokens()).ok_or(InvalidTokens)?;
                 let var_id = tcx.get_diagnostic_item(var_name).ok_or(InvalidTokens)?;
                 contract.variant = Some(var_id);
             }
